+++ spec/files/fail.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+exit Integer(ARGV.fetch(0, '1'))